hide tracks: optionally remove from playlists

+++ src/library/trackcollection.cpp
@@ -282,6 +282,9 @@ bool TrackCollection::hideTracks(const QList<TrackId>& trackIds) {
         }
     }
 
+    // TODO show list with checkboxes for playlists to leave unchanged
+    // use QListWidget, see WTrackMenu::slotRemoveFromDisk()
+    bool removeFromplaylists = false;
     if (!allPlaylistIds.isEmpty()) {
          QStringList playlistNames;
          playlistNames.reserve(allPlaylistIds.count());
@@ -289,21 +292,20 @@ bool TrackCollection::hideTracks(const QList<TrackId>& trackIds) {
              playlistNames.append(m_playlistDao.getPlaylistName(playlistId));
          }
 
-         QString playlistNamesSection =
-                 "\n\n\"" %
+         const QString playlistNamesSection =
+                 "\n\n\"" % // prepend linebreak
                  playlistNames.join("\"\n\"") %
-                 "\"\n\n";
+                 "\"\n\n"; // append linebreak
 
-         if (QMessageBox::question(
-                 nullptr,
+         QMessageBox::StandardButton btn = QMessageBox::question(nullptr,
                  tr("Hiding tracks"),
                  tr("The selected tracks are in the following playlists:"
-                     "%1"
-                     "Hiding them will remove them from these playlists. Continue?")
+                    "%1"
+                    "Do you want to remove them from these playlists. Continue?")
                          .arg(playlistNamesSection),
-                 QMessageBox::Ok | QMessageBox::Cancel) != QMessageBox::Ok) {
-             return false;
-         }
+                 QMessageBox::Yes | QMessageBox::No,
+                 QMessageBox::No);
+         removeFromplaylists = btn == QMessageBox::Yes;
      }
 
     // Transactional
@@ -318,7 +320,9 @@ bool TrackCollection::hideTracks(const QList<TrackId>& trackIds) {
         return false;
     }
 
-    m_playlistDao.removeTracksFromPlaylists(trackIds);
+    if (removeFromplaylists) {
+        m_playlistDao.removeTracksFromPlaylists(trackIds);
+    }
 
     // Post-processing
     // TODO(XXX): Move signals from TrackDAO to TrackCollection
